requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
python-jose[cryptography]
